# Fix mul!(y, A::CompositeMap, x::AbstractMatrix)