--- a/src/composition.jl
+++ b/src/composition.jl
@@ -129,19 +129,23 @@ LinearAlgebra.adjoint(A::CompositeMap{T}) where {T} =
 # comparison of CompositeMap objects
 Base.:(==)(A::CompositeMap, B::CompositeMap) = (eltype(A) == eltype(B) && A.maps == B.maps)
 
-# multiplication with vectors
+# multiplication with vectors/matrices
 _unsafe_mul!(y::AbstractVecOrMat, A::CompositeMap, x::AbstractVector) =
     _compositemul!(y, A, x)
+_unsafe_mul!(y::AbstractMatrix, A::CompositeMap, x::AbstractMatrix) =
+    _compositemul!(y, A, x)
+
 function _compositemul!(y::AbstractVecOrMat,
                         A::CompositeMap{<:Any,<:Tuple{LinearMap}},
-                        x::AbstractVector,
+                        x::AbstractVecOrMat,
                         source = nothing,
                         dest = nothing)
     return _unsafe_mul!(y, A.maps[1], x)
 end
 function _compositemul!(y::AbstractVecOrMat,
-                        A::CompositeMap{<:Any,<:Tuple{LinearMap,LinearMap}}, x::AbstractVector,
-                        source = similar(y, size(A.maps[1], 1)),
+                        A::CompositeMap{<:Any,<:Tuple{LinearMap,LinearMap}},
+                        x::AbstractVecOrMat,
+                        source = similar(y, (size(A.maps[1],1), size(x)[2:end]...)),
                         dest = nothing)
     _unsafe_mul!(source, A.maps[1], x)
     _unsafe_mul!(y, A.maps[2], source)
@@ -149,20 +153,15 @@ function _compositemul!(y::AbstractVecOrMat,
 end
 function _compositemul!(y::AbstractVecOrMat,
                         A::CompositeMap,
-                        x::AbstractVector,
-                        source = similar(y, size(A.maps[1], 1)),
-                        dest = similar(y, size(A.maps[2], 1)))
+                        x::AbstractVecOrMat,
+                        source = similar(y, (size(A.maps[1],1), size(x)[2:end]...)),
+                        dest = similar(y, (size(A.maps[2],1), size(x)[2:end]...)))
     N = length(A.maps)
     _unsafe_mul!(source, A.maps[1], x)
     for n in 2:N-1
-        try
-            resize!(dest, size(A.maps[n], 1))
-        catch err
-            if err == ErrorException("cannot resize array with shared data")
-                dest = similar(y, size(A.maps[n], 1))
-            else
-                rethrow(err)
-            end
+        sz = (size(A.maps[n],1), size(x)[2:end]...)
+        if size(dest) != sz
+            dest = similar(y, sz)
         end
         _unsafe_mul!(dest, A.maps[n], source)
         dest, source = source, dest # alternate dest and source