Performance improvement by avoiding fallback to _generic_mapmat_mul! for
mul!(y, A::CompositeMap, x::AbstractMatrix).

Benchmark:
```
using LinearMaps
using LinearAlgebra
using BenchmarkTools
N = 1000;
A = LinearMaps.CompositeMap{Float64}((LinearMap(randn(N,N)),));
B = randn(N,N);
C = randn(N,N);
X = randn(N,N);
Y = zeros(N,N);
AB = A*B;
ABC = AB*C;

# Before patch
julia> @btime mul!($Y,$A,$X);
  74.417 ms (0 allocations: 0 bytes)

julia> @btime mul!($Y,$AB,$X);
  293.908 ms (1000 allocations: 7.75 MiB)

julia> @btime mul!($Y,$ABC,$X);
  626.289 ms (2000 allocations: 15.50 MiB)


# After patch
julia> @btime mul!($Y,$A,$X);
  11.512 ms (0 allocations: 0 bytes)

julia> @btime mul!($Y,$AB,$X);
  26.461 ms (2 allocations: 7.63 MiB)

julia> @btime mul!($Y,$ABC,$X);
  39.346 ms (4 allocations: 15.26 MiB)
```

Note that I change the source/dest allocation strategy for `CompositeMap`s with >= 3 maps.
Previously, a `resize!` was performed in a try-catch. I think there is no general benefit of a `resize!` here, since it might copy the data if resizing to a bigger array. Furthermore, `resize!` is clumsy to use in the matrix case.
So, I changed to a simpler strategy of only reusing the array if it's exactly the same size.

If I've missed some benefit of using `resize!` I can revert to the previous behavior and update the pull request.