@dkarrasch 2021-03-15
I don't think `resize!` copies data, but I'm not sure. It does some `ccall` eventually, which is... above my paygrade. To play it safe, how about having a vector and a matrix version of the multi-map case?