@rasmushenningsson 2021-03-15
`resize!` does (often) copy data if you change to a larger size vector, because it promises that the existing elements will be kept, and it's unlikely that the array can grow in place. So a new block of memory has to be allocated, which means the data has to be copied to that block.
You can see that the underlying pointer changes below when we `resize!` to a larger array, but not to a smaller.

```
julia> x = zeros(2);

julia> pointer(x)
Ptr{Float64} @0x00007fed2758f6b0

julia> resize!(x,4)
4-element Array{Float64,1}:
 0.0
 0.0
 9.6e-322
 8.55e-322

julia> pointer(x)
Ptr{Float64} @0x00007fed27846a70

julia> resize!(x,2);

julia> pointer(x)
Ptr{Float64} @0x00007fed27846a70
```

But I've updated the code to keep the old solution for the vector case. No reason to change it really - unless I want to spend some time benchmarking different scenarios. :)