@dkarrasch 2021-03-16
That's interesting. I thought it wouldn't because
```julia
y = x = zeros(2)
resize!(x, 5)
y === x # true
```
but I'm not familiar enough with the details to tell what's really going on under the hood.