@dkarrasch 2021-03-16
For the matrix case, I believe this is reasonable enough. In our previous `_generic_mapmat_mul!` branch, we would allocate the intermediate vectors for each column to which we apply the `CompositeMap`, so it ends up allocating basically as much.

Thanks @rasmushenningsson for another nice contribution.